[NFCI] Rename ambiguous use of Client.
Client was ambiguous. It could easily refer to the client running as a browser
(the DomTerm) or a client being managed by the DomServer (like Kawa). The later
use is consistent with X11 applications, but confusing in general.

This commit does not touch any references to the HTML and JS. Those can come in
a separate commit.

Also, I have avoided touching the following in WebTerminal.java:

    jsWebTerminal.setMember("jclient", backend);

Since I didn't want to mess with any of the JS for now.