Experimental with-content-value branch.
The idea is to use span nodes whose class includes 'with-content-value'
(and whose 'content' is provided by a 'value' attribute)
for non-selectable text.
This is initially intended for spacing in input-line elements,
to "hide" identation and such, from selection and cick-moves-mouse.
This seems to mostly work (breakLines2 needs more work),
but it seems to add more complication that it is worth.