Fix parsing of OSC/DCS text parameter to avoid hanging
This reverts handling of text parameter to similar to how it was done
before the byte-parser change: On incomplete text parameter, save it
in a special field rather than the general-purpose _deferredBytes.
That way the receivedCount is incremented and can be confirmed before hanging.
This fixes GitHub issue #94 "Hangs on (almost) any image in browser"