In startOutput, trim down excess blank input lines
This is for "shell integration" and applies after multi-line input
as edited by (say) GNU readline.  If on Enter there are fewer input lines,
then we may have undesirable input styling for part of the output.
Clean this up when we get a start-output command.
(This is not a problem with current fish, but is with current bash/readline.)