Change search for clipboard-reading helper
The libclipboard library is no longer linked with the backend application,
but (if configured) is used to create a small 'dt-libclip' helper binary.
This should reduce the risk that the domterm executable fails to
start because it copied to where libraries (like -lxcb) may be missing.
Also new get_clipboard_command function tries a row of alternative
help commands.