Change DOM-representation of TAB characters
Just uses spaces, but wrapped in a special span.
This fixes line-rewrap, ad generally is more compatible with xterm.
Using attribute dt-tab allows converting back to tab on text-copy.

An alternative to use a content-value attribute but that would be more
work to get working plus selections skip the content-value element.