Avoid "unbreaking" lines unnecessarily.
This fixes a problem caused by readline and similar editors.
Assume 80-column lines and we backspace from 81 columns to 80.
Readline assumes the cursor will stay on the empty overflow line,
but if we re-break the line, we remove the overflow line,
and the cursor will end up in the wrong place.  Instead, assume
the previous lines are still valid, so we don't need to unbreak them.