Improve line-breaking of partial lines.
I.e. when a line already has soft line-breaks, avoid re-breaking.

This isn't just a performance issue, since some applications
(including readline input) may add an extra empty line if the
logical line is an exact multiple of the line length, and if we re-break
we may mess that up.