Try harder to keep connection and session numbers the same.
(1) When allocating a new connection number, avoid the number of an existing
(typically detached) session, even if the connection number is unused.
(2) When allocating a new session number, avoid in-use connection numbers.
(3) When the first session in a main window is detached, we keep the
connection (for window-level requests), but we re-number it so the old
connection number is available if the session is later attached.