Support intial terminal in iframe (DomTerm.useIFrame == 2)
This means creating a separate WebSocket for the top-level.
If we can have the same window-number for both top-level window
and the initial terminal window, the server-side needs some changes.

Setting useIFrame to 2 but keeping layoutInitAlways false mostly works,
including creatng a new sub-window from the comment line.
However, creating a 2nd terminal using (e.g.) C-S-A C-S-Left
causes the window to close for unknown reason.  So for now
we don't support this combination.