New comprehensive framewith for window/pane scale (zoom)
The termOptions field was moved from Terminal to PaneInfo,
to allow options (settings) on a pane which uses a sub-window.
Backend notifies main-window of changed local settings;
the mainWindow resends to the sub-window Terminal if needed.
Moved globalOptions from Terminal to DomTerm and renamed to globalSettings.
Add paneInfo property to Terminal.